refactor(image): 이미지 다이얼로그 공용 컴포넌트로 분리 (#597)
