+++ src/shared/ui/image/index.ts
@@ -1,1 +1,2 @@
 export { HorizontalImageAlbum } from "./HorizontalImageAlbum";
+export { ImageDialog } from "./ImageDialog";
